pyproject.toml
[tool.poetry]
name = "pepes"
version = "0.1.0"
description = "This website aims to store as many pepes as possible"
authors = ["amtlib <grzegorz.andrzej.pach@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.3"
django-rest-framework = "^0.1.0"
Pillow = "^7.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
